+++ tests/PowerUtils.GuardClauses.Validations.Tests/GuardClausesTests/GuardValidationFloatExtensionsTests.cs
@@ -365,4 +365,128 @@ public class GuardValidationFloatExtensionsTests
             "INVALID"
         );
     }
+
+    [Fact]
+    public void IfOutOfRange_InRange_Valid()
+    {
+        // Arrange
+        var quantity = 45.45741f;
+
+
+        // Act
+        var act = Guard.Validate.IfOutOfRange(quantity, 10, 50);
+
+
+        // Assert
+        act.Should()
+            .Be(quantity);
+    }
+
+    [Fact]
+    public void IfOutOfRange_SmallNumber_Exception()
+    {
+        // Arrange
+        var quantity = 5.45741f;
+
+
+        // Act
+        var act = Record.Exception(() => Guard.Validate.IfOutOfRange(quantity, 10, 50));
+
+
+        // Assert
+        act.Validate<PropertyException>(
+            HttpStatusCode.BadRequest,
+            nameof(quantity),
+            "MIN:10"
+        );
+    }
+
+    [Fact]
+    public void IfOutOfRange_BigNumber_Exception()
+    {
+        // Arrange
+        var quantity = 55.45741f;
+
+
+        // Act
+        var act = Record.Exception(() => Guard.Validate.IfOutOfRange(quantity, 10, 50));
+
+
+        // Assert
+        act.Validate<PropertyException>(
+            HttpStatusCode.BadRequest,
+            nameof(quantity),
+            "MAX:50"
+        );
+    }
+
+    [Fact]
+    public void IfOutOfRangeNullable_NULL_Valid()
+    {
+        // Arrange
+        float? quantity = null;
+
+
+        // Act
+        var act = Guard.Validate.IfOutOfRange(quantity, 10, 50);
+
+
+        // Assert
+        act.Should()
+            .Be(quantity);
+    }
+
+    [Fact]
+    public void IfOutOfRangeNullable_InRange_Valid()
+    {
+        // Arrange
+        float? quantity = 45.45741f;
+        
+
+        // Act
+        var act = Guard.Validate.IfOutOfRange(quantity, 10, 50);
+
+
+        // Assert
+        act.Should()
+            .Be(quantity);
+    }
+
+    [Fact]
+    public void IfOutOfRangeNullable_SmallNumber_Exception()
+    {
+        // Arrange
+        float? quantity = 5.45741f;
+        
+
+        // Act
+        var act = Record.Exception(() => Guard.Validate.IfOutOfRange(quantity, 10, 50));
+
+
+        // Assert
+        act.Validate<PropertyException>(
+            HttpStatusCode.BadRequest,
+            nameof(quantity),
+            "MIN:10"
+        );
+    }
+
+    [Fact]
+    public void IfOutOfRangeNullable_BigNumber_Exception()
+    {
+        // Arrange
+        float? quantity = 55.45741f;
+        
+
+        // Act
+        var act = Record.Exception(() => Guard.Validate.IfOutOfRange(quantity, 10, 50));
+
+
+        // Assert
+        act.Validate<PropertyException>(
+            HttpStatusCode.BadRequest,
+            nameof(quantity),
+            "MAX:50"
+        );
+    }
 }