+++ tests/PowerUtils.GuardClauses.Validations.Tests/GuardClausesTests/GuardValidationUShortExtensionsTests.cs
@@ -329,4 +329,128 @@ public class GuardValidationUShortExtensionsTests
         // Assert
         act.Validate<PropertyException>(HttpStatusCode.BadRequest, nameof(quantity), "INVALID");
     }
+
+    [Fact]
+    public void IfOutOfRange_InRange_Valid()
+    {
+        // Arrange
+        ushort quantity = 45;
+
+
+        // Act
+        var act = Guard.Validate.IfOutOfRange(quantity, 10, 50);
+
+
+        // Assert
+        act.Should()
+            .Be(quantity);
+    }
+
+    [Fact]
+    public void IfOutOfRange_SmallNumber_Exception()
+    {
+        // Arrange
+        ushort quantity = 5;
+
+
+        // Act
+        var act = Record.Exception(() => Guard.Validate.IfOutOfRange(quantity, 10, 50));
+
+
+        // Assert
+        act.Validate<PropertyException>(
+            HttpStatusCode.BadRequest,
+            nameof(quantity),
+            "MIN:10"
+        );
+    }
+
+    [Fact]
+    public void IfOutOfRange_BigNumber_Exception()
+    {
+        // Arrange
+        ushort quantity = 55;
+
+
+        // Act
+        var act = Record.Exception(() => Guard.Validate.IfOutOfRange(quantity, 10, 50));
+
+
+        // Assert
+        act.Validate<PropertyException>(
+            HttpStatusCode.BadRequest,
+            nameof(quantity),
+            "MAX:50"
+        );
+    }
+
+    [Fact]
+    public void IfOutOfRangeNullable_NULL_Valid()
+    {
+        // Arrange
+        ushort? quantity = null;
+
+
+        // Act
+        var act = Guard.Validate.IfOutOfRange(quantity, 10, 50);
+
+
+        // Assert
+        act.Should()
+            .Be(quantity);
+    }
+
+    [Fact]
+    public void IfOutOfRangeNullable_InRange_Valid()
+    {
+        // Arrange
+        ushort? quantity = 45;
+
+
+        // Act
+        var act = Guard.Validate.IfOutOfRange(quantity, 10, 50);
+
+
+        // Assert
+        act.Should()
+            .Be(quantity);
+    }
+
+    [Fact]
+    public void IfOutOfRangeNullable_SmallNumber_Exception()
+    {
+        // Arrange
+        ushort? quantity = 5;
+
+
+        // Act
+        var act = Record.Exception(() => Guard.Validate.IfOutOfRange(quantity, 10, 50));
+
+
+        // Assert
+        act.Validate<PropertyException>(
+            HttpStatusCode.BadRequest,
+            nameof(quantity),
+            "MIN:10"
+        );
+    }
+
+    [Fact]
+    public void IfOutOfRangeNullable_BigNumber_Exception()
+    {
+        // Arrange
+        ushort? quantity = 55;
+
+
+        // Act
+        var act = Record.Exception(() => Guard.Validate.IfOutOfRange(quantity, 10, 50));
+
+
+        // Assert
+        act.Validate<PropertyException>(
+            HttpStatusCode.BadRequest,
+            nameof(quantity),
+            "MAX:50"
+        );
+    }
 }